refactor: replace exported styled components with EventRowItem component
Replace bare StyledEventRowItemColumn/StyledEventRowItemAction exports
with a proper EventRowItem component using a variant prop. Extract the
shared props interface into EventRowDynamicComponent.types.ts. This
follows the one-export-per-file convention and keeps styled components
internal to the component file.
