+++ packages/twenty-front/src/modules/activities/timeline-activities/rows/activity/components/EventRowActivity.tsx
@@ -1,11 +1,8 @@
 import { styled } from '@linaria/react';
 import { t } from '@lingui/core/macro';
 
-import {
-  type EventRowDynamicComponentProps,
-  StyledEventRowItemAction,
-  StyledEventRowItemColumn,
-} from '@/activities/timeline-activities/rows/components/EventRowBase';
+import { type EventRowDynamicComponentProps } from '@/activities/timeline-activities/rows/components/EventRowDynamicComponent.types';
+import { EventRowItem } from '@/activities/timeline-activities/rows/components/EventRowItem';
 import { isTimelineActivityWithLinkedRecord } from '@/activities/timeline-activities/types/TimelineActivity';
 import { useOpenRecordInSidePanel } from '@/side-panel/hooks/useOpenRecordInSidePanel';
 import { type CoreObjectNameSingular } from 'twenty-shared/types';
@@ -102,10 +99,10 @@ export const EventRowActivity = ({
     <StyledEventRow>
       <StyledRowContainer>
         <StyledRow>
-          <StyledEventRowItemColumn>{authorFullName}</StyledEventRowItemColumn>
-          <StyledEventRowItemAction>
+          <EventRowItem>{authorFullName}</EventRowItem>
+          <EventRowItem variant="action">
             {t`${eventAction} a related ${eventObject}`}
-          </StyledEventRowItemAction>
+          </EventRowItem>
           <StyledLinkedActivity
             onClick={() =>
               openRecordInSidePanel({
