+++ packages/twenty-front/src/modules/activities/timeline-activities/rows/components/EventRowDynamicComponent.types.ts
@@ -1,0 +1,11 @@
+import { type TimelineActivity } from '@/activities/timeline-activities/types/TimelineActivity';
+import { type EnrichedObjectMetadataItem } from '@/object-metadata/types/EnrichedObjectMetadataItem';
+
+export interface EventRowDynamicComponentProps {
+  labelIdentifierValue: string;
+  event: TimelineActivity;
+  mainObjectMetadataItem: EnrichedObjectMetadataItem;
+  linkedObjectMetadataItem: EnrichedObjectMetadataItem | null;
+  authorFullName: string;
+  createdAt?: string;
+}
