+++ packages/twenty-front/src/modules/activities/timeline-activities/rows/main-object/components/EventRowMainObject.tsx
@@ -1,8 +1,5 @@
-import {
-  type EventRowDynamicComponentProps,
-  StyledEventRowItemAction,
-  StyledEventRowItemColumn,
-} from '@/activities/timeline-activities/rows/components/EventRowBase';
+import { type EventRowDynamicComponentProps } from '@/activities/timeline-activities/rows/components/EventRowDynamicComponent.types';
+import { EventRowItem } from '@/activities/timeline-activities/rows/components/EventRowItem';
 import { EventRowMainObjectUpdated } from '@/activities/timeline-activities/rows/main-object/components/EventRowMainObjectUpdated';
 import { styled } from '@linaria/react';
 import { t } from '@lingui/core/macro';
@@ -54,15 +51,15 @@ export const EventRowMainObject = ({
         <StyledMainContainer>
           <StyledRowContainer>
             <StyledRow>
-              <StyledEventRowItemColumn>
+              <EventRowItem>
                 {labelIdentifierValue}
-              </StyledEventRowItemColumn>
-              <StyledEventRowItemAction>
+              </EventRowItem>
+              <EventRowItem variant="action">
                 {t`was created by`}
-              </StyledEventRowItemAction>
-              <StyledEventRowItemColumn>
+              </EventRowItem>
+              <EventRowItem>
                 {authorFullName}
-              </StyledEventRowItemColumn>
+              </EventRowItem>
             </StyledRow>
             <StyledItemTitleDate>{createdAt}</StyledItemTitleDate>
           </StyledRowContainer>
@@ -85,15 +82,15 @@ export const EventRowMainObject = ({
         <StyledMainContainer>
           <StyledRowContainer>
             <StyledRow>
-              <StyledEventRowItemColumn>
+              <EventRowItem>
                 {labelIdentifierValue}
-              </StyledEventRowItemColumn>
-              <StyledEventRowItemAction>
+              </EventRowItem>
+              <EventRowItem variant="action">
                 {t`was deleted by`}
-              </StyledEventRowItemAction>
-              <StyledEventRowItemColumn>
+              </EventRowItem>
+              <EventRowItem>
                 {authorFullName}
-              </StyledEventRowItemColumn>
+              </EventRowItem>
             </StyledRow>
             <StyledItemTitleDate>{createdAt}</StyledItemTitleDate>
           </StyledRowContainer>
@@ -105,15 +102,15 @@ export const EventRowMainObject = ({
         <StyledMainContainer>
           <StyledRowContainer>
             <StyledRow>
-              <StyledEventRowItemColumn>
+              <EventRowItem>
                 {labelIdentifierValue}
-              </StyledEventRowItemColumn>
-              <StyledEventRowItemAction>
+              </EventRowItem>
+              <EventRowItem variant="action">
                 {t`was restored by`}
-              </StyledEventRowItemAction>
-              <StyledEventRowItemColumn>
+              </EventRowItem>
+              <EventRowItem>
                 {authorFullName}
-              </StyledEventRowItemColumn>
+              </EventRowItem>
             </StyledRow>
             <StyledItemTitleDate>{createdAt}</StyledItemTitleDate>
           </StyledRowContainer>
